# backend/openapi.merged.yaml
openapi: 3.0.3
info:
  title: Kodeme Backend API
  version: 1.0.0
paths:
...
  /calendars:
    get:
      summary: List tenant calendars
      tags:
        - Calendars
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of calendars
          content:
            application/json:
              schema:
                type: object
                properties:
                  calendars:
                    type: array
                    items:
                      $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create a tenant calendar
      tags:
        - Calendars
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orgId
                - userId
                - data
              properties:
                orgId:
                  type: string
                userId:
                  type: string
                data:
                  $ref: '#/components/schemas/CalendarInput'
      responses:
        '200':
          description: Created calendar
          content:
            application/json:
              schema:
                type: object
                properties:
                  calendar:
                    $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /calendars/{id}:
    get:
      summary: Get a tenant calendar (authorized)
      tags:
        - Calendars
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: orgId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Calendar found
          content:
            application/json:
              schema:
                type: object
                properties:
                  calendar:
                    $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Update a tenant calendar
      tags:
        - Calendars
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orgId
                - updates
              properties:
                orgId:
                  type: string
                updates:
                  type: object
                  properties:
                    name:
                      type: string
                    address:
                      type: string
                    location:
                      type: string
                    timezone:
                      type: string
      responses:
        '200':
          description: Updated calendar
          content:
            application/json:
              schema:
                type: object
                properties:
                  calendar:
                    $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete a tenant calendar
      tags:
        - Calendars
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: orgId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /calendars/{id}/unsafe:
    get:
      summary: Get a calendar without org ownership checks (unsafe)
      tags:
        - Calendars
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Calendar or null when not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  calendar:
                    $ref: '#/components/schemas/Calendar'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - action not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    Calendar:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        address:
          type: string
        location:
          type: string
        timezone:
          type: string
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
    CalendarInput:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        location:
          type: string
        timezone:
          type: string
...